Finish max load tests

+++ openreview/arr/process/max_load_process.py
@@ -1,4 +1,4 @@
-def process(client, note, invitation):
+def process(client, edit, invitation):
     import time
     domain = client.get_group(invitation.domain)
     venue_id = domain.id
@@ -10,28 +10,33 @@ def process(client, note, invitation):
     SAC_ID = domain.content['senior_area_chairs_id']['value']
     AC_ID = domain.content['area_chairs_id']['value']
     REV_ID = domain.content['reviewers_id']['value']
-    user = note.signatures[0]
+    user = edit.signatures[0]
 
     edge_readers = [CONFERENCE_ID]
-    if '/Reviewers' in invitation.id:
-        edge_readers += [SAC_ID, AC_ID]
-        role = domain.content['reviewers_id']['value']
-    elif '/Area_Chairs' in invitation.id:
-        edge_readers += [AC_ID]
-        role = domain.content['area_chairs_id']['value']
-    elif '/Senior_Area_Chairs' in invitation.id:
-        role = domain.content['senior_area_chairs_id']['value']
-    else:
-        raise openreview.OpenReviewException('Invalid role for Custom Max Papers edges')
+    inv_role = invitation.id.split('/')[-3]
+    role = None
+    for venue_role in [SAC_ID, AC_ID, REV_ID]:
+      if f"/{inv_role}" in venue_role:
+        role = venue_role
+        if venue_role == AC_ID:
+          edge_readers += [SAC_ID]
+        elif venue_role == REV_ID:
+          edge_readers += [SAC_ID, AC_ID]
+
+    if not role:
+      raise openreview.OpenReviewException('Invalid role for Custom Max Papers edges')
+
     edge_readers += [user]
 
     CUSTOM_MAX_PAPERS_ID = f"{role}/-/Custom_Max_Papers"
     
 
     client.delete_edges(
-          invitation=CUSTOM_MAX_PAPERS_ID,
-          head=user,
-          tail=role
+      invitation=CUSTOM_MAX_PAPERS_ID,
+      head=role,
+      tail=user,
+      wait_to_finish=True,
+      soft_delete=True
     )
   
     client.post_edge(
@@ -42,11 +47,11 @@ def process(client, note, invitation):
         signatures=[CONFERENCE_ID],
         head=role,
         tail=user,
-        weight=int(note.content['maximum_load'])
+        weight=int(edit.note.content['maximum_load']['value'])
       )
     )
 
-    if note.ddate:
+    if edit.note.ddate:
         client.delete_edges(
           invitation=CUSTOM_MAX_PAPERS_ID,
           head=role,
